# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentiment Analysis API",
    "description": "Entraînement et prédiction pour l'analyse de sentiments FR/EN",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/chat/models": {
      "get": {
        "tags": [
          "Chat IA (v1)"
        ],
        "summary": "List Llm Models",
        "description": "Liste les modèles LLM installés sur le serveur (sélecteur du chat).",
        "operationId": "list_llm_models_api_v1_chat_models_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/ai": {
      "post": {
        "tags": [
          "Chat IA (v1)"
        ],
        "summary": "Ai Chat",
        "description": "Chat streaming SSE (thinking_delta / delta / [DONE]) ; erreurs\npré-stream traduites en 502/504 (enveloppe v1).",
        "operationId": "ai_chat_api_v1_chat_ai_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/classifiers/{name}": {
      "get": {
        "tags": [
          "Classifiers (v1)"
        ],
        "summary": "Get Classifier",
        "description": "Instantané d'UN classifieur (public — parité).",
        "operationId": "get_classifier_api_v1_classifiers__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classifiers/{name}/predict": {
      "post": {
        "tags": [
          "Classifiers (v1)"
        ],
        "summary": "Predict Classifier",
        "description": "Prédiction d'un classifieur sur une liste de textes (ordre préservé).",
        "operationId": "predict_classifier_api_v1_classifiers__name__predict_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifierPredictRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassifierPredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classifiers/{name}/reload": {
      "post": {
        "tags": [
          "Classifiers (v1)"
        ],
        "summary": "Reload Classifier",
        "description": "Recharge le modèle actif d'un classifieur depuis le disque.",
        "operationId": "reload_classifier_api_v1_classifiers__name__reload_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessageIn": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessageIn"
      },
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message",
            "description": "Message de l'utilisateur."
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageIn"
            },
            "type": "array",
            "title": "History",
            "default": []
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Nom du modèle LLM Ollama à utiliser (sélecteur du chat). Absent ou vide : modèle par défaut de la configuration serveur."
          },
          "enable_thinking": {
            "type": "boolean",
            "title": "Enable Thinking",
            "description": "Mode « Réflexion » : l'agent raisonne avant de répondre et la trace est diffusée via les événements SSE thinking_delta. Désactivé par défaut (comportement historique inchangé).",
            "default": false
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Session de conversation (app/legacy/core/session_store) où journaliser l'échange ; absent : aucune persistance côté serveur."
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "ClassifierPredictRequest": {
        "properties": {
          "texts": {
            "items": {
              "type": "string",
              "maxLength": 10000
            },
            "type": "array",
            "maxItems": 256,
            "minItems": 1,
            "title": "Texts"
          }
        },
        "type": "object",
        "required": [
          "texts"
        ],
        "title": "ClassifierPredictRequest"
      },
      "ClassifierPredictResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ClassifierPrediction"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "ClassifierPredictResponse"
      },
      "ClassifierPrediction": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "probabilities": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probabilities"
          }
        },
        "type": "object",
        "required": [
          "text",
          "label",
          "confidence"
        ],
        "title": "ClassifierPrediction"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}